requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.3.2
openpyxl==3.1.2
python-multipart==0.0.9
